+++ ZenUnitTests/Random/RandomRunFilter.cpp
@@ -9,7 +9,7 @@ namespace ZenUnit
       RunFilter randomRunFilter;
       randomRunFilter.testClassNamePattern = ZenUnit::Random<std::string>();
       randomRunFilter.testNamePattern = ZenUnit::Random<std::string>();
-      randomRunFilter.testCaseNumber = ZenUnit::Random<unsigned>();
+      randomRunFilter.testCaseNumber = ZenUnit::Random<size_t>();
       return randomRunFilter;
    }
 }
